refactor(execution-policy): narrow max_orders_per_chunk to u32 (DEFI-2743)
`Executor::run_once` casts the chunk size to `usize`, and Rust canisters
target wasm32 where `usize` is 32-bit. Drop the 32 unused bits at the
type level: `NonZeroU64` → `NonZeroU32` on the `ExecutionPolicy` field,
`u32` on `InitArg.max_orders_per_chunk` and `UpgradeArg.max_orders_per_chunk`,
`nat32` in Candid, `Option<u32>` in the snapshot CBOR. Refresh the
canbench init-arg hex and the worst-case-event-size constants
accordingly.

Wire-format break is acceptable: the canister isn't deployed
productively (per the resolved sibling thread on PR #90).

+++ canister/src/state/audit/tests.rs
@@ -84,7 +84,7 @@ impl Scenario {
 
     fn with_upgrade_execution_policy(
         mut self,
-        max_orders_per_chunk: u64,
+        max_orders_per_chunk: u32,
         instruction_budget: u64,
     ) -> Self {
         self.state.set_execution_policy(
